requirements.txt
django==1.11.6
django-bootstrap3

##使用postgresql
##yum install -y python-devel postgresql-devel
#psycopg2

##使用mysql
##yum -y install python-devel mysql-devel
#mysql-python

https://github.com/darklow/django-suit/tarball/v2

channels==1.1.8
asgi-redis==1.4.3
daphne==1.4.2

django_redis
hiredis


docker
docker-compose

pycrypto==2.6.1
paramiko==2.3.1

python-magic

## 登陆时增加otp验证
django_otp
qrcode
pillow

# rsa

# gunicorn
# futures


# djangorestframework

elasticsearch

# pyzabbix

#python2
selectors2

